returning pk of todo for image entity -switched datatype of pks from integer to long as roomdb wasnt recognising int as a return type and to maintain consistency across the app

+++ app/src/main/java/com/team2/todo/data/entities/SubTodo.kt
@@ -18,8 +18,8 @@ import java.time.LocalDateTime
 )
 data class SubTodo(
     @PrimaryKey(autoGenerate = true)
-    val subTodoId: Int,
-    val todoId: Int,
+    val subTodoId: Long,
+    val todoId: Long,
     val title: String?,
     val description: String?,
     /* added type converter for complex date object */
